fix: correctly count FIFO I/O events in edge cases
For reasons that are not entirely clear to me, it seems that `call`
instructions in the `.sched.adb` files can have `oprand_edges` that are
not actually arguments/inputs of the call; specifically, the `ap_fifo`
calls can have numbers of `oprand_edges` that do not correspond to the
number of reads/writes actually being performed by that call. This
commit fixes that by ignoring any `oprand_edge` whose sink is not the
`call` instruction being considered.