pin llvmlite to version 0.43
This is a hack. Our LLVM distribution is (a fork of) LLVM 15, and the
latest version of llvmlite (0.44) upgrades its LLVM library version to
LLVM 15. Because conda injects its lib directory into the rpath of the
clang we build before the actual intended lib directory of our own
distribution, if we use llvmlite=0.44 which depends on libllvm15, then
at runtime, our clang picks up the wrong libLLVM-15.so. These leads to
problems, the obvious one being that the @llvm.part.* instrinsics we
added back in are no longer implemented, leading to testbench link
failures.

llvmlite=0.43 depends on libllvm14, sidestepping the issue (and is still
newer than the version we originally developed with).

This was not a fun issue to figure out.